chore(typedoc-plugin-appium): reorganize the tests
many of these were actually E2E tests (and slow)
